1539126: Don't use expires: never
This also fixes a bug when metrics.yaml files contain a date for expires.
PyYAML will convert those dates into datetime objects, which can't be validated
by JSON schema (which doesn't understand datetimes).  We want that always to
be strings, so we use a workaround to get PyYAML to not generate datetimes.