Make labels from Labeled types sets again
These labels were recently changed to lists to ensure the labels
ordering from `metrics.yaml` was kept in the markdown generated
documentation. That had the unwanted effect of changing the type
of the labels in the Kotlin and Swift generated codes.

To fix this, a new property `ordered_labels` was added to the
Labeled types which is a list. Keeping the actual `labels` property
a set.