Add conditional Foundation import to swift.jinja2 template
Xcode started complaining about this while trying to consume Glean via the AppServices megazord (which is required for A-S components to use the Glean Rust bindings on iOS). I purposefully made this conditional to protect against the cases where Foundation may not be importable/available, which I _think_ should keep this working in our sample app which doesn't consume Glean from rust-components-swift.