Update JS encoder to properly serialize Rate metrics
The rate metric previously was not supported in newer versions of the
glean_parser for Glean.js. The rate metric is implemented in Glean.js,
glean_parser just didn't have the functionality in place to properly
serialize the Rate metric data, likely why this was set to skip the
last time this was touched.

The JS encoder now pulls in similar logic from the swift & kotlin
encoders to make sure that the objects that are passed from the Rate
metric are serialized.