Allow unit only for the quantity metric
This is implemented the same way time_unit is implemented for others:
The specific metric class takes the argument, not the generic one.

Because the schema enforces `unit` to be required for `quantity` we can
just `pop` it with the guarantee it's there.

The error message for other metrics if they do set `unit` is less nice:

    Metric.__init__() got an unexpected keyword argument 'unit'

But that's in line with what `time_unit` also does.